Moved and updated functions
