+++ src/commands/study/ChemInfo.ts
@@ -3,7 +3,13 @@ import BaseCommand, { type DiscordChatInputCommandInteraction } from "../../regi
 import type { DiscordClient } from "@/registry/DiscordClient";
 import type { ChemInfo, SynonymInfo, ExperimentalInfo } from "@/utils/apis/cheminfo";
 import { Logger } from "@discordforge/logger";
-import { metals, atomicNumberToSymbol } from "@/data.ts";
+import { 
+    fetchChemicalData,
+    fetchChemicalSynonyms,
+    determineBonding,
+    getExperimentalProperties,
+    formatFormula 
+} from "@/utils/apis/cheminfo";
 
 export default class ChemInfoCommand extends BaseCommand {
     constructor() {
@@ -99,128 +105,4 @@ export default class ChemInfoCommand extends BaseCommand {
             await interaction.editReply("Invalid formula or name (an error occurred)");
         }
     }
-}
-
-
-async function fetchChemicalData(chemical: string, namespace: string):
-    Promise<ChemInfo.APIResponse | ChemInfo.APIErrorResponse> {
-        const res = (await fetch(
-            `https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/${namespace}/${chemical}/JSON`
-            ).then((res) => res.json())) as ChemInfo.APIResponse | ChemInfo.APIErrorResponse;
-        return res;
-}
-
-
-async function fetchChemicalSynonyms(
-    cid: number
-): Promise<SynonymInfo.APIResponse> {
-    const res = (await fetch(
-        `https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/${cid}/synonyms/JSON`
-        ).then((res) => res.json())) as SynonymInfo.APIResponse;
-    return res;
-}
-
-
-async function determineBonding(
-    compound: ChemInfo.PCCompound, 
-    is_element: boolean, 
-    is_ion: boolean
-): Promise<string | null> {
-    const atoms = compound.atoms.element.map(atomicNumber => atomicNumberToSymbol[atomicNumber] || '');
-
-    if (is_ion) return "Can form ionic bonds";
-    if (!is_element && atoms.some(atom => metals.includes(atom))) return "Ionic bonds";
-    if (atoms.some(atom => metals.includes(atom))) return "Metallic bonds (Metal)";
-    if (!is_element) return "Covalent bonds";
-
-    return null;
-}
-
-
-async function getExperimentalProperties(
-    cid: number
-): Promise<ExperimentalInfo.ExperimentalProperties> {
-    try {
-        const url = `https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/${cid}/JSON/`;
-        
-        const response = await fetch(url);
-        const data = await response.json() as ExperimentalInfo.ApiResponse;
-
-        let mainSection = null;
-
-        for (const section of data.Record.Section) {
-            if (section.TOCHeading === "Chemical and Physical Properties") {
-                mainSection = section;
-                break;
-            }
-        }
-
-        if (!mainSection) {
-            return { description: null, color: null };
-        }
-
-        let experimentalProperties = null;
-
-        for (const subsection of mainSection.Section) {
-            if (subsection.TOCHeading === "Experimental Properties") {
-                experimentalProperties = subsection;
-                break;
-            }
-        }
-
-        if (!experimentalProperties) {
-            return { description: null, color: null };
-        }
-
-        let description = null;
-        let color = null;
-
-        for (const property of experimentalProperties.Section) {
-            if (property.TOCHeading === "Physical Description") {
-                description = property.Information[0].Value.StringWithMarkup[0].String;
-                color = property.Information[1].Value.StringWithMarkup[0].String;
-                
-                if (color) {
-                    color = color.split(";")[0];
-                }
-                break;
-            }
-        }
-
-        return { description, color };
-    } catch (error) {
-        return { description: null, color: null };
-      }
-}
-
-
-async function formatFormula(formula: string): Promise<string> {
-    const translations: { [key: string]: string } = {
-        "+1": "⁺¹",
-        "+2": "⁺²",
-        "+3": "⁺³",
-        "+4": "⁺⁴",
-        "+5": "⁺⁵",
-        "+6": "⁺⁶",
-        "+7": "⁺⁷",
-        "-1": "⁻¹",
-        "-2": "⁻²",
-        "-3": "⁻³",
-        "-4": "⁻⁴",
-        "-5": "⁻⁵",
-        "-6": "⁻⁶",
-        "-7": "⁻⁷",
-        "0": "₀",
-        "1": "₁",
-        "2": "₂",
-        "3": "₃",
-        "4": "₄",
-        "5": "₅",
-        "6": "₆",
-        "7": "₇",
-        "8": "₈",
-        "9": "₉",
-    };
-    const regex_rule = new RegExp(Object.keys(translations).map(key => key.replace(/[+-]/g, '\\$&')).join("|"), "g");
-    return formula.replace(regex_rule, match => translations[match]);
 }
